Added PCGRandom.raw_seed(size) -> string
Added a new module function ::raw_seed that generates a ruby string
composed of randomly generated bytes.

This deprecated the existing pcg_new_seed_bigmul() internal function
which will no longer be used and is hence removed from the source as it is
superseded by pcg_new_seed_bytestr()

The current technique is to source the desired number of bytes from a
platform entropy source and represent it as a ruby string which is the output
of ::raw_seed. ::new_seed goes a step further and unpacks this 16B sequence
into a single biginter via what is effectively:

PCGRandom.new_seed -> PCGRandom.raw_seed.unpack("Q*")[0] -> Bignum