Updates to compilation strategy and preliminary gem release. There still may be some kind of memory issues
# PCGRandom

PCGRandom is now defined as a class!

## PCGRandom.new_seed(size = 128)

* PCGRandom.new_seed is now a singleton method of class PCGRandom, and also takes
an optional parameter `size`, an even number which determines the size of the
collective seed to generate. This raises an ArgumentError for odd numbers

For example:
> PCGRandom.new_seed 1
ArgumentError: Seed size must be even! Found 1
from (pry):3:in `new_seed'
> PCGRandom.new_seed 16
=> 50277733213615332021408984479931419745536
> PCGRandom.new_seed 18
=> 3979171113203830563784127101378611462326448640
> PCGRandom.new_seed 40
=> 4141971354660884649132630595236286361156917961892615239640546237371781661805312
> PCGRandom.new_seed 64
=> 9109964008685802937263770162100209408873975481367774050864525418302360391842509742861349886316874679287815962808130048

## Internals
* fallback_entropy_getbytes() no longer appears in the code since
entropy_getbytes() already references this function as **it's own fallback**

## Build

* dir_config() is run before adding stuff, which actually makes sense and
explains why the compiled .so file was so bad where pcg_random wasn't
already installed

Bumped version to 0.1.2